nii.get_fdata() will cast the ndarray to float64, but we always use float32 or less, so I changed it to nii.get_fdata(dtype=np.float32) (it is not possible to cast directly to 8 or 16 bit w/ get_fdata)

+++ Heifets_lab_scripts/max.py
@@ -2,6 +2,7 @@
 
 import sys
 import nibabel as nib
+import numpy as np
 
 file_path = sys.argv[1]
 
@@ -9,7 +10,7 @@ file_path = sys.argv[1]
 nii_img = nib.load(file_path)
 
 # Get the data from the file
-data = nii_img.get_fdata()
+data = nii_img.get_fdata(dtype=np.float32)
 
 # Find the maximum intensity value in the data
 max_intensity = int(data.max())
